Split test_merge.py into metadata / glyph_data / pipeline files
Move shared paths and merge harnesses into conftest.py, then group
the 30+ test classes by intent:
- test_metadata.py: nameID / OFL / version / settings text
- test_glyph_data.py: outlines / metrics / hints / GSUB / GPOS / CFF
- test_pipeline.py: CID, base-only, WOFF2, packaging, output dir
